Update `NextTemplate` module with new components
