+++ src/modules/index.ts
@@ -1,2 +1,1 @@
-export { default as NextTemplate } from './NextTemplate/NextTemplate';
 export { default as SpaceX } from './SpaceX/SpaceX';
