+++ src/pages/index.tsx
@@ -1,23 +1,64 @@
 import type { NextPage } from 'next';
-import Head from 'next/head';
+import { useTranslation } from 'react-i18next';
 
-import { Container, Footer } from '@common/components';
-import { NextTemplate, SpaceX } from '@src/modules';
+import { VStack } from '@chakra-ui/react';
+import { AppContainer } from '@src/common/components';
+import {
+  Title,
+  Description,
+  NextInformation,
+} from '@modules/NextTemplate/sections';
 
 const Home: NextPage = () => {
+  const { t } = useTranslation('nextTemplate');
+  const head = {
+    title: 'Create Next App',
+    metas: [
+      {
+        name: 'description',
+        content: 'Generated by create next app',
+      },
+    ],
+    links: [
+      {
+        rel: 'icon',
+        href: '/favicon.ico',
+      },
+    ],
+  };
+
+  const cards: Card[] = [
+    {
+      href: 'https://nextjs.org/docs',
+      title: t('cards.0.title'),
+      description: t('cards.0.description'),
+    },
+    {
+      href: 'https://nextjs.org/learn',
+      title: t('cards.1.title'),
+      description: t('cards.1.description'),
+    },
+    {
+      href: 'https://github.com/vercel/next.js/tree/canary/examples',
+      title: t('cards.2.title'),
+      description: t('cards.2.description'),
+    },
+    {
+      href: 'https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app',
+      title: t('cards.3.title'),
+      description: t('cards.3.description'),
+    },
+  ];
+
   return (
-    <Container>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+    <AppContainer head={head}>
+      <VStack flex={1} justify="center" align="center" gap="2rem">
+        <Title />
 
-      <NextTemplate />
-      <SpaceX />
-
-      <Footer />
-    </Container>
+        <Description />
+        <NextInformation cards={cards} />
+      </VStack>
+    </AppContainer>
   );
 };
